ai/requirements.txt
absl-py==1.4.0
astunparse==1.6.3
blis==0.7.9
cachetools==5.3.1
catalogue==2.0.8
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
confection==0.1.0
contourpy==1.1.0
cycler==0.11.0
cymem==2.0.7
distlib==0.3.7
filelock==3.12.4
flatbuffers==23.5.26
fonttools==4.40.0
gast==0.4.0
google-auth==2.21.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.56.0
h5py==3.9.0
idna==3.4
importlib-metadata==6.7.0
importlib-resources==5.12.0
Jinja2==3.1.2
joblib==1.2.0
keras==2.13.1
keras-facenet==0.3.2
keras-models==0.0.7
kiwisolver==1.4.4
langcodes==3.3.0
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.7.1
mtcnn==0.1.1
murmurhash==1.0.9
nltk==3.8.1
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.8.0.74
opt-einsum==3.3.0
packaging==23.1
pathlib==1.0.1
pathy==0.10.2
Pillow==10.0.0
platformdirs==3.10.0
preshed==3.0.8
protobuf==4.23.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.11
pyparsing==3.1.0
python-dateutil==2.8.2
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.0
scikit-multilearn==0.2.0
scipy==1.10.1
six==1.16.0
sklearn==0.0.post5
smart-open==6.3.0
spacy==3.5.4
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
thinc==8.1.10
threadpoolctl==3.1.0
tqdm==4.64.1
typer==0.9.0
typing_extensions==4.5.0
urllib3==1.26.14
virtualenv==20.24.5
wasabi==1.1.2
Werkzeug==2.3.6
wrapt==1.15.0
zipp==3.15.0
